Add functionality to the button in the move passwords promo card
This CL adds logic to the action button in the move passwords promo
card. When pressed, this button will show move_passwords_dialog.
When move_passwords_dialog move button is pressed, it checks whether
the promo card triggered it, and if so it checks if all of the
local passwords were moved and closes the promo card.

Additionally passwords_section checks for butter requirements, and if
the are lost when the move passwords promo card is displayed, it hides
the promo card.

This CL also fixes small UX issue where promo card close button was too
close to the edge, by adding margin that the UX suggested.

Low-Coverage-Reason: HARD_TO_TEST
Bug: 1503146
Change-Id: Ie94f214b40ea4dea2f08e569775e18db13283b2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5171797
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Karol Sygiet <sygiet@google.com>
Cr-Commit-Position: refs/heads/main@{#1247097}