[ios] Fix hangs during shutdown if many sync-ing tabs
The main practical difference is that some observer functions are
invoked a bit later:
1. WebStateObserver::WebStateDestroyed()
2. WebStatePolicyDecider::ResetWebState()

With ~85 overrides for the first, and although many of them appear to
be trivial (e.g. just removes the observer), it is hard to be sure
about undesired side effects.

That said, one of them, IOSChromeLocalSessionEventRouter::Observer,
is a clear beneficiary: the fact that WebStateList gets trimmed at
once, and before WebStateDestroyed() is invoked, avoids quadratic
runtime behavior in sync code, which iterates the entire WebStateList
every time WebStateDestroyed() is invoked.

Change-Id: Ib84f8391fe9f374719111fa04cfc0cf45423aeb8
Bug: 1510661
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5119913
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247131}