[blink] Correct CanvasResourceRasterSI SI usage flags with killswitch
The adjustment of SI usage flags in
CanvasResourceResourceRasterSharedImage for OOP-R being enabled is
currently a bit wonky:

- Even if OOP-R is not enabled, RASTER usage should be passed. However,
  it's currently passed only with OOP-R enabled.

This CL makes this correction, protected by a killswitch as it can have
an impact on the SI backing used. Theoretically, this change should be
unproblematic: It should result only in a "more correct" backing being
selected if anything.

However, I want to protect against any weird edge-case interactions with
service-side code that might need to be shaken out and corrected.

Bug: 1518427
Change-Id: I2e4627d00dc5198c64ce538a084aa556abde1bb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5197786
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247203}