[IdleTimeout] Part 1 of EG Tests for IdleTimeout and IdleTimeoutActions
This is the first part of the tests for IdleTimeout policies egtests.
This mainly tests the cases for showing the dialog and the snackbar.
More tests will be written to test specific actions.

The cases tested and their expectations are as follows:

- `Continue using Chrome` button clicked: no actions run, the snackbar is now shown and the dialog is not dismissed
- The dialog counts down 30 seconds without user interaction: dialog is dismissed, actions run and the snackbar is shown.
- The dialog is shown then the app is backgrounded: on re-foreground, the actions run and the dialog is not shown. The snackbar shows after actions run.
- Multiple windows: The dialog shows in one of the windows
- Multiple windows and the window showing the dialog is closed: the dialog should transfer to the other window still opened.

The CL also fixes and improves a few bugs that affects some tests and improve consistency of the variables that the service has access to.

Bug: b/301676922
Change-Id: Ie59bb90498f7adc6202bb66f1c9cbfc8c167b229
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5148170
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Vincent Boisselle <vincb@google.com>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Cr-Commit-Position: refs/heads/main@{#1247223}