Remove ChromeAppsWebViewPermissiveBehaviorAllowed policy
The planned milestone for the removal of this escape hatch policy has
arrived. We mark the policy as no longer supported and remove the code
that implements it.

We also remove the chrome://flags entry which controls the same
behaviour. However, we leave the underlying base::Feature for a follow
up CL to remove.

Bug: 1261928, 1508572
Change-Id: I9e30d8523348b96bbebecf7b7701948c9652064f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5185285
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: W. James Maclean <wjmaclean@chromium.org>
Reviewed-by: Victor Gabriel Savu <vsavu@google.com>
Cr-Commit-Position: refs/heads/main@{#1247249}