Remove GrabViewSnapshot from render_frame_host_impl_browsertest
The async version is now wired up, and the sync version is deprecated.
The code to call the async version is already there so switch to
always using it.

Note that this test was accidentally passing on Linux and Android
because it didn’t test the result bool and assumed the sync screenshot
call succeeded when in fact it was a no-op. This CL marks the whole
test as skipped, but switches in the correct API so that the test
owners can take their time and make it work.

Bug: 1517390
Change-Id: I7d06eac5ed66824a591eb3bd06944442a345ba7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5190329
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: William Liu <liuwilliam@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247293}