Mark the shelf as opaque for occlusion
On display scale change, exo windows are briefly the wrong size due to
async configuration update. If they are maximized and the display is
zoomed in, if the shelf cannot occlude what is below it, exo windows can
have their occlusion state changed temporarily. For example, if the top
window is resized to be smaller, the below maximized window will be
unoccluded where the shelf is until it is also resized.

So, let the shelf occlude things below it.

Bug: 1399478
Test: tested in a follow up CL as part of occlusion work
Change-Id: I6da60ae41604e9ec9f8922419e942f65f3145bf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5119061
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Auto-Submit: Eliot Courtney <edcourtney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247330}