Clip occludees to the root window
On display scale change, exo windows are briefly the wrong size due to
async configuration updates. If those windows are maximized and the
display is zoomed in, the exo windows will temporarily be larger than
the root window. Then, if the top window is resized to be smaller, the
below maximized window will become unoccluded in the portion which is
outside the root window and be incorrectly marked as visible. So, this
CL clips regions to the root window bounds.

Bug: 1399478
Test: added unit test + local testing
Change-Id: Ic078e57f914a180474f76a44bceca571fd8ffd53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5119062
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247332}