Sort windows from top to bottom for display update
This CL sorts windows from top to bottom when performing a display
configuration update (like display scale change). One reason to do this
is when the windows get larger, if we resize the topmost one first, it
will occlude continue to anything below it while they are also resized
to be larger. We pre-order DFS the window 🌲 from bottom-most to
top-most and collect the windows to avoid quadratic lookup.

In general, there is no best order which optimises non-atomic changes,
but top to bottom is one of the more reasonable ones.

Bug: 1509181
Test: added unit test + local testing
Change-Id: I1e263068b430334ac6b112495cb5881f0ae13e8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5118683
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247334}