Start Android Stylus handwriting after edit element is focused
Android Stylus handwriting is getting started before edit element
is focused because of which Platform tries to associate it with
last InputConnection and tries to commit there but recognized text
should be committed to edit element in which stylus handwriting
is happening.

Start stylus handwriting after edit element is focused so that
InputConnection is current focused element.

Bug: 1512519, 1503673
Change-Id: I038326e3a04e9274a0c0579eba8df163986599a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5124965
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247427}