[Passwords] Fix showing suggestions on fields with manual generation
When a password field is focused, AutofillAgent queries
PasswordGenerationAgent for suggestions, and if no generation is
required, it queries PasswordAutofillAgent.

As per comments in the header,
PasswordGenerationAgent::ShowPasswordGenerationSuggestions should
return true only when generation suggestions are offered.
But in case when generation is triggered manually, and then the
value is erased by the user, the method keeps returning true, even
though the generation suggestions are not shown anymore.

Fixing the bug by allowing the method to return false, then
AutofillAgent is able to query PasswordAutofillAgent.

Bug: 1495325
Change-Id: I621d26cb234c6fe77ee4a09201338eaa0ee831e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5198206
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247453}