Avoid recovery if the device settings load succeeded on boot
On device boot, the device settings are loaded synchronously before
checking the EnrollmentConfig. This allows us to check again if the
enrollment recovery is marked properly.

In this change we add an extra-check on boot after the device is
marked as requiring recovery to check again on device settings.

The UMA will allow us to identify if such cases happen and also to
see if serial_number missing cases are happening.

BUG=b:231677312
TEST=Manual test on device

Change-Id: Ic093d6e6f83e8d304ce199649031db49257d17aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5172670
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247472}