[Autofill] Simplify PopulateFormFieldsAndFrames()
This CL simplifies PopulateFormFieldsAndFrames(): instead of an
out-parameter, the function now returns a FormData.

As a result, calling CreateFormWithoutFieldsAndFrames() from
PopulateFormFieldsAndFrames() seems like the best option again.
This CL makes this call an anonymous lambda call in order to
make it very clear on which parameters it depends.

This CL is a no-op.

Bug: 1007974
Change-Id: I0a6fc1c2004c5f00a6c72327619410093a95d28d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5199003
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1247473}