[Fix] Can't dismiss remote admin notification when there are no local accounts.
We missed to cover the case when there are no local accounts presents
and the OOBE flow must be completed Hence, the dialog was forced to stay
open with the remote activity notification screen. Fixed this issue by
opening the previous screen present if no user is present.

Bug: b/319635930
Change-Id: I55458134fadad475ca706cd63cd1e413427499dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5190769
Reviewed-by: Jeroen Dhollander <jeroendh@google.com>
Commit-Queue: Ashutosh Singhal <macinashutosh@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/main@{#1247475}