[DBSC] Fix bug in overriding a session
Gaia is currently using a constant session id. The current
implementation of `RegisterNewBoundSession`:
- Saves the new session params
- If there is an existing session it clears it from storage.
- If session id and storage are the same, this will also clear the new
  session.

In practice, this can happen if the user is redirected to sign in and a
new session is started. With Gaia using the same session id, this will
result in the bound session being terminated on next startup (session
cleared from storage) and the user signed out.

Bug: b/320292767
Change-Id: I8724adc4105f7ba84c18af0962f04aa22c42a140
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5198850
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1247490}