[Profiles][CR2023] Fix padding for select avatar in profile creation
This CL fixes a bug where unnecessary scrollbar were shown in the select
avatar page in the local profile creation dialog with Chrome Refresh
2023 (ChromeWebuiRefresh2023 flag enabled).

The height of the action-button changed from 32px to 36px with CR2023.
The padding was adjusted to fix the total height of the page.

Screenshots with ChromeWebuiRefresh2023 enabled and disabled:
https://crbug.com/1514893#c5

Fixed: 1514893
Change-Id: I547b4b4a70540930cddfaa8e793071c420de1cf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5200169
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Gabriel Oliveira <gabolvr@google.com>
Cr-Commit-Position: refs/heads/main@{#1247543}