Fix VariationsService et al being recreated on shutdown.
If shutdown has started and MetricsServicesManager and the
objects it owns (including VariationsService) have been
destroyed already, return nullptr from their getters, instead
of recreating them.

Note: Call sites are already expected to check for null and
the one known call site that was calling us on shutdown does
already. If any other call site does this but doesn't check
for null, we'll notice it as a result of a null pointer deref
crash and can fix it as a follow-up.

The iOS code is also updated to have this defensive logic,
even if this issue was not observed on iOS so far.

Bug: 1517706
Change-Id: Ifa13e241842b41ee12bad4a9c653c38465b0f6ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5197856
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247551}