Reduce the range of GestureTime metric
The metric records how long it takes to perform a stylus gesture. However, because of the size of the buckets, the majority (70%) of data was being recorded in the 0-10ms bucket which wasn't helpful. This change should create more buckets between 0 and 10ms which should more accurately show the average time for gestures.
https://uma.googleplex.com/p/chrome/timeline_v2?sid=72efbe30334198ebf0273d5ac6fc83e7

Change-Id: Ie345d8aeba55c8293130c3fd9fa3330492db5d90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5192455
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Commit-Queue: Alex Mitra <alexmitra@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247555}