Remove ContentLayerClient::PaintableRegion()
Reasons:
1. All implementations of ContentLayerClient just returned
   gfx::Rect(layer_bounds).
2. cc::RecordSource and cc::RasterSource can get the total bounds
   of visual rects of paint operations from cc::DisplayItemList
   cheaply, without needing the client to tell cc with the function.

This CL doesn't change any visible behavior.

In the next steps for crbug.com/1517714, we will use
DisplayItemList::bounds() which is added in this CL returning
the total bounds of the painted results to determine the tiling
area to avoid unnecessary tiles in areas that are just for
hit-testing.

Bug: 1517714
Change-Id: Ia1d57fea15ec47de43e00e7bb034fc9a8581406f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5190710
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247623}