[mullet m2] Fix having no label when triggering field is Address Line 1
When the triggering field contains Address Line 1, the code also
considered Address Line 1 as a label candidate, which ended up actually
adding it as a label. Our logic prevents having the same label as the
main text, so the label was completely removed.

Even if the profile had a country, the country is only added as a label
as a last resort. The new test has two profiles which only differ by
country. This test prevents this issue.


Bug: 1459990
Change-Id: I575788b9f03827f2bf8a29d3164d9af7562f5fd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5192456
Commit-Queue: Theo Cristea <theocristea@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Bruno Braga <brunobraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1247639}