Ship SpeculationRulesImplicitSource.
It's a minor, presumably uncontroversial, change, so we should be good
to just ship this. This can be reverted in code (reverting this CL) or
remotely (via Finch) if that turns out not to be the case.

PSA: https://groups.google.com/a/chromium.org/g/blink-dev/c/lyY10E6yLsY/m/9vFyWFHNBgAJ

Bug: 1517696
Change-Id: I01b6c33364193f89ec5c95697d819ce7f6808ba7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5189773
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247643}