Emit form similarity check metric in more instances.
This CL also adds form similarity check result emission in
OnAskForValuesToFill and on OnFocusOnFormField. The motivation behind
these additions is to understand what the impact of changing
FormDataAndroid::SimilarFormAs would be on the parts of Android
Autofill that are not related to prefill requests.

Low-Coverage-Reason: COVERAGE_UNDERREPORTED
Bug: 1513847
Change-Id: Idf3377458310df73087a651a64b25095443d84a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5198366
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1247649}