Pass color provider to CSSParser::ParseSystemColor
This CL is a follow up on CL:5085762. In this CL, we plumb color
providers that exist in Blink::Page to the CSSParser::ParseSystemColor.

The CL starts by adding an extra ColorProvider argument to the
ParseSystemColor method and updates all of its call sites.
In order to get the color provider from the Page via the document,
we introduce a GetColorProvider method in BaseRenderingContext2D
class that returns a color provider from the document.

This CL does not use the color providers to get the system colors yet.
The next steps would involve updating StyleColor::ColorFromKeyword to
take in a color provider, using the passed color provider in
LayoutTheme::SystemColor and remove the SystemColorFromNativeTheme
method from LayoutTheme.

Bug: 1430181
Change-Id: Ia91774198bb74f685b94cbceab2b6bd21c1d8748
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5121572
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247695}