SM Debug: Loads events that had been recorded before debug page was loaded.
Events that have been recorded lose much of the information that is
human readable. To simplify the implementation, events that are recorded
while the page is open and events that are loaded use the same
interface. The webpage is notified of new events once they have been
fully processed instead of before.

NO_BUG

Change-Id: Ibdb0318a0ff189d355665b190c2864740e965d62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5086287
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Commit-Queue: Andrew Bregger <andrewbregger@google.com>
Cr-Commit-Position: refs/heads/main@{#1247727}