Make the event type check for d&d more explicit
A user may release a mouse button that is not a part of d&d operation,
and it can still cancel the drag and drop. Instead, explicitly check
the left button (or other buttons), to decide if the drag and drop
should be canceled.

A few extra changes
* Moved ShellClientData from shell_unittest.cc for reuse
* Added wl_pointer support to TestClient
* changed set_data so that it returns the pointer of moved object.
  I added DestroyData as template can't take nullptr any more.
* Added generic button support to EventGenerator

Bug: 1513068
Test: covered by unittest
Low-Coverage-Reason: 　TRIVIAL_CHANGE  logging code.
Change-Id: Icf148e785cc51d222886dc744285b3a923dbdcb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5142213
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247730}